Move drink association classes out of models into their own module
