+++ beers/models/__init__.py
@@ -1,11 +1,13 @@
 from beers.models.models import (
-    Contest, Contest_Beer, Contest_Bonus, Contest_Brewery,
-    Contest_Checkin, Contest_Player
+    Contest, Contest_Checkin, Contest_Player
 )
 
 from beers.models.player import Player
 from beers.models.drinks import Beer, Brewery
 from beers.models.checkin import Unvalidated_Checkin
+from beers.models.associations import (
+    Contest_Beer, Contest_Bonus, Contest_Brewery,
+)
 
 __all__ = [
     'Beer', 'Brewery', 'Contest', 'Contest_Beer', 'Contest_Bonus', 'Contest_Brewery',
