+++ beers/models/models.py
@@ -26,8 +26,6 @@ class ContestManager(models.Manager):
         contest_player = contest.add_player(creator)
         return contest
 
-
-
 class Contest(models.Model):
     "Represents a contest"
 
@@ -55,6 +53,7 @@ class Contest(models.Model):
         return contest_player
 
     def add_beer(self, beer, point_value=1):
+        from .associations import Contest_Beer
         """Adds a beer into the contest"""
         beer = Contest_Beer(contest=self, beer=beer,
                             beer_name=beer.name,
@@ -76,6 +75,7 @@ class Contest(models.Model):
         will get 3 for drinking it and also have the challenger lose 3 points
         up to a maximum of 12 points lost
         """
+        from .associations import Contest_Beer
         beer = Contest_Beer(contest=self, 
                             beer=beer,
                             challenger=challenger,
@@ -90,6 +90,7 @@ class Contest(models.Model):
 
     def add_brewery(self, brewery, point_value=1):
         """Adds a brewery to the contest"""
+        from .associations import Contest_Brewery
         brewery = Contest_Brewery(contest=self, brewery=brewery,
                                   brewery_name=brewery.name,
                                   point_value=point_value,)
@@ -120,6 +121,7 @@ class Contest(models.Model):
         Checks if a hash tag is already in use for this contest. Returns matching 
         bonus or None if no bonus exists.
         """
+        from .associations import Contest_Bonus
         return Contest_Bonus.objects.filter(contest=self, 
                                             hashtags__contains=[hash_tag]).first()
 
@@ -130,6 +132,7 @@ class Contest(models.Model):
         hash_tags can be a plain string, a comma-separated string, or an array of 
         strings
         """
+        from .associations import Contest_Bonus
         tag_list = self.__clean_hash_tags(hash_tags)
         tag_matches = [(tag, self.__using_hash_tag(tag)) for tag in tag_list]
         tag_conflict = list(filter(lambda t: t[1] is not None, tag_matches))
@@ -161,6 +164,7 @@ class Contest(models.Model):
         Gets the list of beers in beer, brewery order. If the player is passed
         in, it adds a checked_into=True to the entry
         """
+        from .associations import Contest_Beer
         if player is None:
             return Contest_Beer.objects.filter(
                 contest=self).order_by('beer_name')
@@ -176,6 +180,7 @@ class Contest(models.Model):
         Gets the list of breweries in name order. If the player is passed
         in, it adds a checked_into=True to the entry
         """
+        from .associations import Contest_Brewery
         if player is None:
             return Contest_Brewery.objects.filter(
                 contest=self).order_by('brewery_name')
@@ -188,26 +193,6 @@ class Contest(models.Model):
 
     def __str__(self):
         return self.name
-
-class Contest_BreweryManager(models.Manager):
-
-    def link(self, contest, brewery, value):
-        return self.create(contest=contest, brewery=brewery,
-                           brewery_name=brewery.name,
-                           point_value=value, total_drank=0,)
-
-class Contest_Brewery(models.Model):
-    contest = models.ForeignKey(Contest, on_delete=models.CASCADE)
-    brewery = models.ForeignKey(Brewery, on_delete=models.CASCADE)
-    brewery_name = models.CharField(max_length=250)
-    point_value = models.IntegerField(default=1)
-    total_visited = models.IntegerField(default=0, 
-			help_text="number of players who drank at this brewery")
-
-    objects = Contest_BreweryManager()
-
-    def __str__(self):
-        return '{}/{}'.format(self.contest.name, self.brewery.name)
 
 class Contest_Player(models.Model):
     """ Links a player's activities relative to a contest
@@ -422,6 +407,7 @@ class Contest_Player(models.Model):
 
         returns Contest_Checkin
         """
+        from .associations import Contest_Bonus
         checkin_time = None
         untappd_checkin = None
         contest_bonus = None
@@ -490,46 +476,6 @@ class Contest_Player(models.Model):
     def __str__(self):
         return "{0}:[Player={1}]".format(self.contest.name, self.user_name)
 
-class Contest_Beer(models.Model):
-    "Represents a many-to-many connection between a beer and a contest"
-
-    contest = models.ForeignKey(Contest, on_delete=models.CASCADE)
-    challenger = models.ForeignKey(Contest_Player,
-                                   on_delete=models.CASCADE,
-                                   default=None,
-                                   null=True,
-                                  )
-    beer = models.ForeignKey(Beer, on_delete=models.CASCADE)
-    beer_name = models.CharField(max_length=250)
-    brewery_name = models.CharField(max_length=250, default='')
-    point_value = models.IntegerField(default=1)
-    challenge_point_loss = models.IntegerField(default=0)
-    max_point_loss = models.IntegerField(default=0)
-    challenge_point_value = models.IntegerField(
-        default=0,
-        help_text='The number of points the challenger gets ' +
-            'for drinking this beer')
-    total_drank = models.IntegerField("number of players who drank this beer")
-
-    class Meta:
-        unique_together = (('contest', 'beer'),)
-
-    def __str__(self):
-        return "{0}/{1}".format(self.beer.name, self.beer.brewery)
-
-class Contest_Bonus(models.Model):
-    """Represents a bonus associated with a particular contest"""
-
-    contest = models.ForeignKey(Contest, on_delete=models.CASCADE)
-    name = models.CharField(max_length=50, default=None, null=False, blank=False,)
-    description = models.CharField(max_length=250, default="", null=True, blank=True,)
-    hashtags = ArrayField(base_field=models.CharField(max_length=30, default=None),
-                          null=True, 
-                          default=None)
-    point_value = models.IntegerField(default=1,)
-
-    class Meta:
-        unique_together = (('contest', 'name'),)
 class Contest_CheckinManager(models.Manager):
     def create_checkin(self, contest_player, contest_beer, checkin_time,
                        untappd_checkin):
@@ -549,6 +495,7 @@ class Contest_CheckinManager(models.Manager):
                            untappd_checkin=untappd_checkin)
 
 class Contest_Checkin(models.Model):
+    from .associations import Contest_Beer, Contest_Bonus, Contest_Brewery
     TRANSACTION_TYPES = (
         ('BE', 'Beer'),
         ('BR', 'Brewery'),
